Remove mock macro
